move function out of component and into its own utils file
